+++ src/colony_print/test/controllers/job.py
@@ -1,0 +1,39 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+import logging
+import unittest
+
+import appier
+
+import colony_print
+
+
+class JobControllerTest(unittest.TestCase):
+    def setUp(self):
+        self.app = colony_print.ColonyPrintApp(level=logging.ERROR)
+
+    def tearDown(self):
+        self.app.unload()
+        adapter = appier.get_adapter()
+        adapter.drop_db()
+
+    def test_list(self):
+        response = self.app.get("/jobs")
+        self.assertEqual(response.code, 403)
+
+    def test_show(self):
+        response = self.app.get("/jobs/name")
+        self.assertEqual(response.code, 403)
+
+    def test_cancel(self):
+        response = self.app.post("/jobs/name/cancel")
+        self.assertEqual(response.code, 403)
+
+    def test_files(self):
+        response = self.app.get("/jobs/name/files")
+        self.assertEqual(response.code, 403)
+
+    def test_file(self):
+        response = self.app.get("/jobs/name/files/file.txt")
+        self.assertEqual(response.code, 403)